Move most x86 specific modules to arch/x86
