+++ src/arch/x86/cpu_freq_sampling.hpp
@@ -6,9 +6,9 @@
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-//
+// 
 //     http://www.apache.org/licenses/LICENSE-2.0
-//
+// 
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -16,36 +16,25 @@
 // limitations under the License.
 
 #pragma once
-#ifndef HW_APIC_HPP
-#define HW_APIC_HPP
+#ifndef X86_CPU_FREQ_SAMPLING_HPP
+#define X86_CPU_FREQ_SAMPLING_HPP
 
-#include <cstdint>
-#include <delegate>
-#include "xapic.hpp"
-#include "x2apic.hpp"
-
-namespace hw
+namespace x86
 {
-  class APIC {
-  public:
-    static IApic& get() noexcept;
+  static const int CPU_FREQUENCY_SAMPLES = 18;
 
-    // enable and disable legacy IRQs
-    static void enable_irq (uint8_t irq);
-    static void disable_irq(uint8_t irq);
+  extern void   reset_cpufreq_sampling();
+  extern double calculate_cpu_frequency();
 
-    static uint8_t get_isr() noexcept {
-      return get().get_isr();
-    }
-    static uint8_t get_irr() noexcept {
-      return get().get_irr();
-    }
-    static void eoi() noexcept {
-      return get().eoi();
-    }
+  /** Proper IRQ-handler for CPU frequency sampling - implemented in interrupts.s
+      @Note 
+      PIT::estimateCPUFrequency() will register- /de-register this as needed */
+  extern "C" void cpu_sampling_irq_handler();
 
-    static void init();
-  };
-}
+  /** CPU frequency sampling. Implemented in hw/cpu_freq_sampling.cpp 
+      @Note this will be automatically called by the oirq-handler  */
+  extern "C" void cpu_sampling_irq_entry();
+
+} //< namespace hw
 
 #endif
