+++ src/arch/x86/pic.cpp
@@ -15,31 +15,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#pragma once
-#ifndef HW_SMP_HPP
-#define HW_SMP_HPP
+#include <hw/ioport.hpp>
+#include "pic.hpp"
 
-#include <cstdint>
-#include <delegate>
-#include <vector>
+namespace x86 {
 
-namespace hw {
+  uint16_t PIC::irq_mask_ {0xFFFF};
 
-  class SMP {
-  public:
-    typedef delegate<void()>  smp_task_func;
-    typedef delegate<void()>  smp_done_func;
+  void PIC::init() noexcept {
 
-    // add tasks that will not necessarily start immediately
-    // use work_signal() to guarantee work starts
-    static void add_task(smp_task_func, smp_done_func);
-    // call this to signal that work is queued up
-    static void work_signal();
-    static std::vector<smp_done_func> get_completed();
+    hw::outb(master_ctrl, icw1 | icw1_icw4_needed);
+    hw::outb(slave_ctrl,  icw1 | icw1_icw4_needed);
+    hw::outb(master_mask, icw2_irq_base_master);
+    hw::outb(slave_mask,  icw2_irq_base_slave);
+    hw::outb(master_mask, icw3_slave_location);
+    hw::outb(slave_mask,  icw3_slave_id);
 
-    static void init();
-  };
+    hw::outb(master_mask, icw4_8086_mode | icw4_auto_eoi);
+    hw::outb(slave_mask,  icw4_8086_mode);  // AEOI-mode only works for master
 
-}
-
-#endif
+    set_intr_mask(irq_mask_);
+  }
+} //< namespace hw
