+++ src/kernel/rtc.cpp
@@ -2,7 +2,6 @@
 
 #include <kernel/os.hpp>
 #include <kernel/timers.hpp>
-#include <hw/cpu.hpp>
 #include <hw/cmos.hpp>
 #include <hertz>
 
@@ -22,20 +21,20 @@ void RTC::init()
 
   // set current timestamp and ticks
   current_time  = cmos::now().to_epoch();
-  current_ticks = hw::CPU::rdtsc();
+  current_ticks = OS::cycles_since_boot();
 
   INFO("RTC", "Enabling regular clock sync with CMOS");
   // every minute recalibrate
   Timers::periodic(seconds(60), seconds(60),
   [] (Timers::id_t) {
     current_time  = cmos::now().to_epoch();
-    current_ticks = hw::CPU::rdtsc();
+    current_ticks = OS::cycles_since_boot();
   });
 }
 
 RTC::timestamp_t RTC::now()
 {
-  auto ticks = hw::CPU::rdtsc() - current_ticks;
+  auto ticks = OS::cycles_since_boot() - current_ticks;
   auto diff  = ticks / Hz(MHz(OS::cpu_freq())).count();
 
   return current_time + diff;
