+++ src/kernel/softreset.cpp
@@ -5,7 +5,7 @@
 #define SOFT_RESET_MAGIC    0xFEE1DEAD
 #define SOFT_RESET_LOCATION 0x7000
 
-namespace hw {
+namespace x86 {
   extern uint32_t apic_timer_get_ticks() noexcept;
   extern void     apic_timer_set_ticks(uint32_t) noexcept;
 }
@@ -39,7 +39,7 @@ void OS::resume_softreset(intptr_t addr)
   
   /// restore known values
   OS::cpu_mhz_ = data->cpu_freq;
-  hw::apic_timer_set_ticks(data->apic_ticks);
+  x86::apic_timer_set_ticks(data->apic_ticks);
 }
 
 extern "C"
@@ -49,7 +49,7 @@ void* __os_store_soft_reset()
   auto* data = (softreset_t*) SOFT_RESET_LOCATION;
   data->checksum    = 0;
   data->cpu_freq    = OS::cpu_freq();
-  data->apic_ticks  = hw::apic_timer_get_ticks();
+  data->apic_ticks  = x86::apic_timer_get_ticks();
   
   uint32_t csum = crc32(data, sizeof(softreset_t));
   data->checksum = csum;
